requirements.txt
absl-py==0.14.1
asgiref==3.4.1
astunparse==1.6.3
attrs==21.2.0
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
clang==5.0
cloudinary==1.28.0
cryptography==35.0.0
cycler==0.10.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
dj3-cloudinary-storage==0.0.6
Django==3.2.8
django-allauth==0.45.0
django-cleanup==5.2.0
django-environ==0.8.1
django-heroku==0.0.0
django-materializecss-form==1.1.17
django-toolbelt==0.0.1
ffmpeg-python==0.2.0
ffpyplayer==4.3.2
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
gunicorn==20.1.0
h5py==3.1.0
idna==3.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
mediapipe==0.8.8
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
opt-einsum==3.3.0
Pillow==8.3.2
protobuf==3.18.1
psycopg2==2.9.2
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-magic==0.4.24
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.2
static3==0.7.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.12.1
